# openapi.json
{"openapi":"3.1.0","info":{"title":"Сириус.Аренда","description":"API для сервиса бронирования пространств","version":"1.0.0"},"paths":{"/rooms":{"post":{"tags":["rooms","rooms"],"summary":"Создание комнаты (только администратор)","operationId":"create_room_rooms_post","parameters":[{"name":"access_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"403":{"description":"Недостаточно прав"},"409":{"description":"Комната с таким названием существует"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["rooms","rooms"],"summary":"Получить все комнаты","operationId":"get_rooms_rooms_get","parameters":[{"name":"min_cap","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Min Cap"}},{"name":"max_cap","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":0},{"type":"null"}],"title":"Max Cap"}},{"name":"equipment","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Equipment"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"},"title":"Response Get Rooms Rooms Get"}}}},"400":{"description":"Минимальная вместимость больше максимальной"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rooms/{room_id}":{"get":{"tags":["rooms","rooms"],"summary":"Получить подробности комнаты по айди","operationId":"get_room_rooms__room_id__get","parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"string","title":"Room Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"404":{"description":"Комната не найдена"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["rooms","rooms"],"summary":"Изменить данные комнаты (только администратор)","operationId":"update_room_rooms__room_id__put","parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"string","title":"Room Id"}},{"name":"access_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"403":{"description":"Недостаточно прав"},"404":{"description":"Комната не найдена"},"409":{"description":"Комната с таким названием существует"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["rooms","rooms"],"summary":"Удалить комнату (только администратор)","operationId":"delete_room_rooms__room_id__delete","parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"string","title":"Room Id"}},{"name":"access_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"}}],"responses":{"204":{"description":"Successful Response"},"403":{"description":"Недостаточно прав"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/bookings":{"post":{"tags":["bookings","bookings"],"summary":"Создание брони","operationId":"create_booking_bookings_post","parameters":[{"name":"access_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Комната не найдена"},"400":{"description":"Неверный формат даты"},"409":{"description":"Пространство уже занято"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["bookings","bookings"],"summary":"Получить все брони","operationId":"get_all_books_bookings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/bookings/{booking_id}":{"delete":{"tags":["bookings","bookings"],"summary":"Отмена брони","operationId":"cancel_booking_bookings__booking_id__delete","parameters":[{"name":"booking_id","in":"path","required":true,"schema":{"type":"string","title":"Booking Id"}},{"name":"access_token","in":"cookie","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Access Token"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Бронь не найдена, либо пользователь её не бронировал"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rooms/{room_id}/bookings":{"get":{"tags":["bookings","bookings"],"summary":"Получить все брони по дате и комнате","operationId":"get_booking_rooms__room_id__bookings_get","parameters":[{"name":"room_id","in":"path","required":true,"schema":{"type":"string","title":"Room Id"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"Неверный формат даты"},"404":{"description":"Броней нет"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"tags":["auth","auth"],"summary":"Регистрация пользователя","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User_url"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"409":{"description":"Человек с таким именем существует"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["auth","auth"],"summary":"Авторизация пользователя","operationId":"login_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Неверный пароль"},"404":{"description":"Не существует такого пользователя"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BookingCreate":{"properties":{"room_id":{"type":"string","title":"Room Id"},"date_start":{"type":"string","title":"Date Start"},"date_end":{"type":"string","title":"Date End"},"status":{"$ref":"#/components/schemas/BookingStatus","default":"активно"}},"type":"object","required":["room_id","date_start","date_end"],"title":"BookingCreate"},"BookingStatus":{"type":"string","enum":["активно","отменено"],"title":"BookingStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Room":{"properties":{"name":{"type":"string","title":"Name"},"capacity":{"type":"integer","title":"Capacity"},"equipment":{"items":{"type":"string"},"type":"array","title":"Equipment"},"id":{"type":"string","title":"Id"}},"type":"object","required":["name","capacity","equipment","id"],"title":"Room"},"RoomCreate":{"properties":{"name":{"type":"string","title":"Name"},"capacity":{"type":"integer","title":"Capacity"},"equipment":{"items":{"type":"string"},"type":"array","title":"Equipment"}},"type":"object","required":["name","capacity","equipment"],"title":"RoomCreate"},"User":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"User"},"User_url":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role","default":"base"}},"type":"object","required":["username","password"],"title":"User_url"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}